Re-organizing the code, ensuring ContactTracing is used in other models, and removing SAMPLE_FROM macros (now a member of Model)
