+++ include/epiworld/tool-bones.hpp
@@ -101,7 +101,7 @@ public:
     ///@}
 
     void set_name(std::string name);
-    std::string get_name() const;
+    virtual std::string get_name() const;
 
     Agent<TSeq> * get_agent();
     int get_id() const;
